Add BSD support to erlang_js
This patch adds .mk files for FreeBSD, NetBSD, OpenBSD and DragonFlyBSD.
It also patches some nspr files so that nspr builds correctly for the
BSDs as well.

The only thing not solved by this commit is the make/gmake issue, which
is also a problem on solaris. On BSD GNU make is called 'gmake' and BSD
make is called 'make'. However rebar is blindly assuming that GNU make
is called 'make', so this is problematic on any non-Linux UNIX flavor.